[vulkan] Pass correct output_padding_arg when creating conv2d contexts
Passing `{}` as `output_padding` results in the following exception

```
c10::Error: expected output_padding to be a single integer value or a list of 2 values to match the convolution dimensions, but got output_padding=[]
```

when calling

```
const auto output_padding =
      expand_param_if_needed(output_padding_arg, "output_padding", 2);
```

in the constructor of `Conv2dPackedContext`. This can be fixed by passing `{0}` instead.

Differential Revision: [D38449709](https://our.internmc.facebook.com/intern/diff/D38449709/)

ghstack-source-id: 163621611
Pull Request resolved: https://github.com/pytorch/pytorch/pull/82861