[ao] Fix per-channel histogram visualization in ModelReportVisualizer
Summary: There was an issue with per-channel visualizations in the
ModelReportVisualizer that in specific scenarios in which there were
only per-channel features for a module, it would fail to specifically
get the channel by channel info.

After digging through the code, the core reason was a for loop that was
enumerating on the `tensor_table` (tensor level info) even in the
scenario in which we only had per-channel info.

This was fixed, and tested in a Bento to ensure expected functionality.

Test Plan: Tested visually

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 2ec2b21db5c04351f4d9360da782b5c10872785e
Pull Request resolved: https://github.com/pytorch/pytorch/pull/82917