[ao] Updated ModelReportVisualizer per-channel line plot
Summary: Before, the line plot for the ModelReportVisualizer used to
plot a different line for each channel. However, for models that have a
lot of channels, this can get really hard to read and parse and doesn't
provide much valuable information.

Now, we just have a single value per module that is the average of the
500 channels.

We also considered plotting 3 lines (a min line, a max line, and an
average line) but the issue was that large outliers could result in one
of the lines completely messing up the scale and the other two not being
visible. As a result, it made sense to do an average and let the user
use the report data to generate the other two if they wished to do so.

This was tested visually in a ipynb notebook

Test Plan: Tested visually in a ipynb notebook

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 7c9e1099c87a3e05ac67b1e2fe744cc634bb08a2
Pull Request resolved: https://github.com/pytorch/pytorch/pull/82918