[Quant][fx][bc-breaking] Integrate BackendConfig with quantization flow (part 2)
Pull Request resolved: https://github.com/pytorch/pytorch/pull/82557


This is part 2 of the effort to replace `backend_config_dict` with
a python config object, a more formal and robust API that leads to
better user experience. This commit integrates the `BackendConfig`
implemented in part 1 (https://github.com/pytorch/pytorch/pull/81469)
with the existing FX graph mode quantization flow.
ghstack-source-id: 163748027

Differential Revision: [D38471932](https://our.internmc.facebook.com/intern/diff/D38471932/)