Fix error_inputs for diag
`args` must be a tuple. It was not a tuple, so the error that was being
checked was actually being raised in the testing suite and not in
torch.diag itself.

This PR changes the error input to check something more reasonable: that
the input must be a matrix or a vector.

Test Plan:
- run tests

ghstack-source-id: 8699be2fceeb48e9a0da58a973a7404c9222530d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/82984